formalize the runner, suite and view in the config helpers
This cleans up some usage of the config helpers in the runner, suite
and view.  The config helpers now formally define each object so
the things mixing in the config helpers don't have to.

Now the runner no longer manually has to define the objects it needs.
I also had each object override the config helpers that define
themselves.  Finally, this renames the run time and rates to include
`formatted_` in the name.  The previous name collided with with
the corresponding names on the suite.  It wasn't causing problems, it
was just bad form.

This is more prep for switching to have the tests accumulate result
data rather than store it as objects in memory.