runner: add `-t` option to run only a single test
This updates assert's base runner to allow running only a single
test.  The idea is to not only be able to isolate failing tests
and work on just them but also to print out assert CLI commands
in fail/error messages to allow you to easily re-run only a single
failing test for quick iterations.

To accomplish this I've added a new cli opt, `-t`, that takes the
file/line of the single test to run.  For example:

```
$ assert -t test/unit/config_tests.rb:61
```

This option configures the file line for the single test and puts
assert in "single test" mode.  Only the test file for this test
will be loaded and only that single test will be run.  If an invalid
file/line is specified, no tests are run.

Some addtional notes:

* `-t` was used b/c you are calling out a test - t for test
* the file line parse logic was updated to be a bit more flexible -
  it not parses even if the file or line parts are omitted