runner: move test randomization to the base runner
This moves the logic/handling for randomizing the test run order
out of the default runner and into the base runner.  I'm essentially
saying this is not something that is intended to be overridden.

This is prep for adding logic to just run just single tests.  I want
to centralize the "baseline" features into the the base runner so
they are always available, not matter what runner is in use.  This
is also prep for adding a custom runner that runs tests in parallel.
It shouldn't have to re-implement this kind of basline behavior.