output the cmd to re-run a result's test in result details
Now that we can run individual tests using the new `-t` option,
a common need is to re-run just a single test that is failing and
we want to focus on.  This updates the result details that are dumped
by the default view to build and include an assert cmd for re-running
just that result's test.

To make this happen, I first needed to track the `test_id` on each
result.  The test id is just the file/line string for the result's
test.

I chose to print the command with no empty lines between it and the
rest of the result details b/c we currently don't put any empty lines
in the result details.  I chose to not ansi color it to help make it
stand out from the rest of the details.  This is similar to how we
output captured output for each result.