switch to applying CLI options last
Previously, ENV settings would be applied last and could therefore
override CLI settings.  It seems strange in hindsight to ever override
something that has been specified at the CLI.  It leads to confusing
and unpredictable behavior as you may not realize and ENV var is
even set.

This switches to always applying CLI settings last so that they
are never overridden.  I also chose to reorder some of the implementation
to match the order b/c that is what was done previously.  I also
updated the README to keep it relevent and matching the implementation.