debug mode: bench showing debug msg immediately; show timing on finish
This addresses issue 256.  On test suites that are large (many test
files to require or complicated env setup in the test helper), debug
mode is only somewhat helpful.  On the longer init steps (ie requiring
the test helper and requiring the test files) assert would just stare
blankly at you while stuff was happening but you didn't know what
it was doing.  Only once the step was complete did you know what it
was doing.  This is not great UX.

This updates the bench logic to pring the debug msg immediately to
show that this step has begun and we are waiting on it.  Then
when the step is done the timing info is displayed as before.  This
lets the developer see that something is going on and start mentally
estimating how long it has taken so far.

This also justifies the output so the timing info is easier to read.
Again, just trying to improve the UX a bit.