stubs: have stub errors use the backtrace of the original caller
Previously, the backtraces of the stub errors would be nested in
the internals of the stubbing implementation.  This is really
confusing b/c you really want the backtrace to point to where the
problem originated so that the user is shown what they need to change.

This updates the stubbing internals to capture the original caller
for any calls that may raise stubbing errors.  This original caller
is then used as the backtrace for all stub errors.

Note: this also makes the newly added `stub_send` raise a `NotStubbedError`
when trying to call the original method on something that hasn't
been stubbed.  This is consistent with the other stub methods and
should have been done in PR 258 where this method was originally
added but was missed.