apply runner seed whether running single or multiple tests
When adding the single test mode back in PR 250, I ended up only
applying the runner seed if you were running in non-single test mode
(ie not using the `-t` option flag).  On the surface this kinda sense
b/c the seed drives the test order randomization. If you are only
running one test you don't need test randomization.

However, the intent is that this seed also drives the pseudo random
generator (`srand`). This impacts the values that are generated by
Assert's factories. We need to always apply the seed info so that you
can re-run a single test and get the same factory values out.  This
update ensures that whether running in single or multi test mode, the
seed value is always applied.