accumulate: rework suite to abstract the tests to run
This is part of switching to accumulating test run and result data.
This is specifically prep for removing the `tests` reader on the
suite.  Since we will be accumulating data, there will be no need
to keep a list of all the tests in memory.

I chose to move the storage of tests as an array into the default
suite as opposed to the base suite.  I don't want to force storing
tests in memory on all suites.  Some suites may choose to store
their tests to run some other way.

I chose to add an `on_test` callback for adding tests to the suite.
This follows the callback pattern used for other custom behavior
(ie `on_result`) and removes needing to access the raw list of
tests outside of the suite.

I chose to add a method to clear the tests to run.  This is prep
for clearing out the tests from memory once they have been run so
they can be garbage collected.

Some other really minor stuff:

* I removed the ansi-related comments.  I'm not sure what I originally
  intended with these but I don't want to do that anymore.
* I also added a check to make sure there are tests to run before
  printing the runner seed info - this fixes a bug I introduced
  in a previous effort.  The seed info would print even if there
  were no tests to run and this was unintended.