accumulate: rework the default view to accumulate results to dump
This implementation no longer relies on the result objects stored
on each test object that was run.  This is part of switching to
no longer storing test objs that were run and the result objs those
tests produced.

This new implmentation accumulates the results that need to be
dumped as results come in.  This has a few advantages over the
previous implementation.  For starters, not needing to store results
means we use less memory in the test suite (which has a big impact
on large test suites).  Also, this decouples the result object
from the default view b/c we store result data custom to the view.
Only the attrs that need displaying will be stored.  Finally, this
only has to sort a subset of all results generated (the ones that
need to be dumped).  This is more performant on large test suites
that generate many results (most of which are just passing and are
not dumped).

Due to these changes all of the suites result handling methods can
be removed.  All result handling is now expected to be done by the
views themselves (since they are the users of the result data).

There are a couple of additional changes that came out of this
effort.  First, I switched the `ansi_styled_message` view helper
to expect a result type symbol instead of taking either a result or
result type.  This simplifies the impelementation and is of no
consequence given we are collecting result types in the views result
data.

Second, this updates how results are sorted when dumped.
The result now knows its `file_line` (the file/line-num of the test
code that produced the result) and this is collected in the view's
result data.  The view sorts result data in reverse file line order.
This is the same effect the previous implementation had, just more
explicit about it.  The main advantage here is that, no matter what
order results come in, they are always outputted in an expected order.
This will be needed if tests are ever run in parallel so I figured
I should go ahead and be explicit about it now.