(hotfix) track `test_file_line` on results; alias `test_id`
This is a minor cleanup in prep for accumulating test data on runs.
`test_file_line` is actually the more appropriate attr name for this
data (its built from a test's `file_line.to_s`) and it should have
been called this originally.  However, `test_id` is a needed attr
as well.  It just so happens that the test file line is used as the
test id.

This adds the more appropriately named attr while keeping the orig
attr intact for previous needs.  This will be used by the coming
accumulate test data on test runs feature as we will store test
run data by the test file line and will need to access that data
as results come in from test runs.

Note: this also removes the legacy `data` method on results.  This
method is no longer being used so no need keeping it around.