accumulate: move in-memory test handling back to base suite
It was premature to assume this behavior was specific to just the
default suite.  Any suite, regardless of how they accumulate test
or result counts, will need to keep the tests in memory as they
need the test code blocks to run.  I discussed the likely need to
do this back in PR 268.

This also removes the tests accessor on the suite and updates the
descriptions of both the suite and default suite.  The tests that
were relying on the tests accessor were updated accordingly to use
the proper test handling methods on the suite.

This is all part of the effort to accumulate test run data instead
of storing off the actualy test objs and the result objs they
produce.