accumulate: properly sort result data by the file name and line num
Previously, the data just stored the file_line string which did not
sort the file line number portion as intended b/c string integers
don't sort like integers.  This switches the storing the file name
and line number of the result independently and using those to
properly sort the results by reverse definition order.

This edge case was missed in PR 269 but was missed.  This is part
of the effort to accumulate test run data instead of storing test
objects.