accumulate: more fixes to result handling and view sorting
This is part of reworking assert to accumulate test run data and
came out of trying to remove the `results` accessor from the test
object (a future effort).  In working on this effort, I broke a
bunch of tests.  I noticed that the test results weren't ordered
as I expected and I started to investigate.

The first thing I noticed was that for the error result backtraces,
the line used for the result file line included extra characters
after the line number (`".../test_tests.rb:123 in some_method"` for
example).  The file line wasn't parsing the backtrace lines
correctly.  So I updated the file line to properly parse backtrace
lines in this form.

Once that was corrected, I noticed that errors that occurred in
Assert's lib source code were being shown first, regardless of the
definition order of the test that generated the error result.  The
first problem was that the result file line used to sort the results
was built off the first filtered backtrace line.  If the backtrace
only contained assert lib lines, the filtered backtrace would be
empty.  So I added a check for empty filtered backtraces and set
the `trace` attr to the first filtered line if present but fall
back to just the first line of the backtrace if not.  Note: the
case where the filtered backtrace is empty is pretty edge - it
should only occur when errors in Assert's test suite come from the
lib source.  This shouldn't occur in any other test suites.

This change got the errors coming from Assert's lib source to now
have a file line but they were still not sorting as expected.  This
was b/c their file line (in Assert's lib source) comes before any
results with a file line in Assert's test source.  This is true of
any test suite - the results generated from non-test code would
always be sorted before results generated from test code,  So this
led me to rework how the view's result data is sorted.  It is now
sorted first by the test's file line and then by the results file
line. This does two things: it ensures all results are jux'd next
to other results from the same test and results are always sorted
by the test that generates them. Within a single test, results will
be sorted by the code file line that generated the result.

Changing the result data sort behavior required me to formally
model not only the test file line, but also the test file name and
test line num.  This matches the result file name and line num
attrs and allows proper sorting.  I also chose to make the test file
line an actual file line object.  The test id attr is still the
string form of the test file line (no change from before).

Note: I also updated the result `inspect` method to include both
the file line and test file line info.  I figured this was helpful
info for determining where a result came from when debugging.