accumulate: prep tests for removing `results` from test objects
This reworks a bunch of the assertion tests and system tests to
not expect a test to know its results.  This is prep for removing
the `results` accessor from tests, which in turn is part of switching
to accumulating test run data.

I chose to break out a test helpers module so I could include it
on all the test contexts.  Since the tests all needed to accumulate
results and tested the results in a similar manner, this commonization
works.

I chose to break this out into a separate PR to remove some noise
from the coming effort where I remove the `results` accessor on
the Test class.