accumulate: stop storing results/result-state on the test objects
This is part of the effort to accumulate test run data and not store
it on the objects themselves.  This stops storing the result objects
on the test that produces them.  Now that we are accumulating the
result data needed for presentation on the views, we don't need to
keep these objects around.  This should result in tests suites
using less memory and hopefully performing better.

This removes not only the results themselves, but all methods that
in some way calculate or store result-state data.  None of this is
needed anymore since the view is accumulating the result data it
needs for presentation.