(hotfix) more test prep for removing `results` from test objects
This is similar to the work done in PR 275 and should have been
included in that effort but was missed.

This reworks the Test tests that assert the Test is producting
captured results to not expect the test itself to provide these
results.  Instead it switches to using the common test helpers
that capture results outside of the test (just like the assertion
tests do).  This is prep for removing the `results` accessor from
tests, which in turn is part of switching to accumulating test run
data.

Like in PR 275, I chose to break this out into a separate commit
to remove noise from the coming effort where I remove the `results`
accessor on the Test class.