(hotfix) even more test prep for removing `results` from test objs
This is similar to the work done in PR 275 and should have been
included in that effort but was missed.

Even more so this should have been included in the previous hotfix
(7eaebab) that addressed reworks to the Test tests that assert the
Test is producing captured results to not expec the test itself to
provide these results.  Like that hotfix, this switches to using
the common test helpers that capture results outside of the test.
This is prep for removing the `results` accessor from tests, which
in turn is part of switching to accumulating test run data.

Like in PR 275 and 7eaebab, I chose to break this out into a sep
commit to remove noise from the coming effort where I remove the
`results` accessor on the Test class.