accumulate: clear out tests from the suite as they are run.
This is a slight optimization related to now accumulating test run
data.  We no longer need to keep references to the tests-to-run
since we are accumulating their meta and result data.  This switches
to clearing out the suite tests-to-run just before the running
starts iterating them.  It also updates the runner to remove its
reference to the test as each test is done running.

Note: I had to update a couple of runner tests where they were
incorrectly relying on tests being available post-run.  This
should have been done back when I added these tests but was missed
b/c we kept the tests around and I just didn't notice at the time.