(hotfix) use array w/ single test path when globbing in single test mode
This fixes a bug that only shows up in modern ruby versions (ie non-1.8.7).
In ruby 1.8.7 string responds to `inject` but in modern versions it
does not.  This was hiding a bug where I wasn't passing an array
to the private `globbed_test_files` method in single test mode.

This fixes the bug which allows single test mode to work in modern
ruby versions.