replace the stubbing API with MuchStub
We have decided to extract the stubbing API to a separate gem,
MuchStub, so that it can be used independently from Assert (ie in
other testing frameworks like Rspec and Minitest).  This removes
all stubbing related logic and switches to bringing in MuchStub
as a dependency.

Note: the main `Assert.{stub|unstub|stub_send|etc}` api is still
available (it just proxies MuchStub now).  There are no backwards
incompatible changes with this.