replace the factory API with MuchFactory
We have decided to extract the factory API to a separate gem,
MuchFactory, so that it can be used independently from Assert (ie
in other testing frameworks like Rspec and Minitest). This removes
all factory related logic and switches to bringing in MuchFactory
as a dependency.

All of the factory api is the same, there are no backwards
incompatible changes with this.

Note: I chose to extend/include MuchFactory onto Assert::Factory
instead of just setting `Assert::Factory = MuchFactory`.  This is
b/c you should be able to further extend or modify Assert::Factory
without modifying MuchFactory.