add Assert::Context::LetDSL
This adds a `let` DSL method for declaring test input values. These
create local methods to instance_eval the block and memoize its
return value.

The benefit of using lets instead of manually defining methods
for test inputs is that lets auto-memoize for you and put the value
declarations at the beginning of the test context definition.

The benefit of using lets instead of e.g. setting an instance
variable in a `setup` block, is that they are late-bound and the
value is only memoized if the let method is called.