add `assert_that` assertion helper; add `Assert::ActualValue`
This enables defining assertions in the form:

```ruby
assert_that(1).equals(1)
```

This form removes the ambiguity of the argument order in the
original assertion helpers and reads more like a formal assertion.
This is just another way to write tests and keeps Assert flexible
and un-imposing on the style that tests are written in.