rewrite Assert's test suite to use `assert_that` assertions
This rewrites all of the tests in Assert to use the `assert_that`
assertion style where possible. This not only provides a rich
system test for `assert_that`, there is no loss of coverage b/c
`assert_that` calls the original assertion helpers.

The main win in switching to this style is that it removes the
argument order ambiguity of expected vs actual values when they
are passed to the assertion helpers.