allow passing a block to `assert_that`
This addresses Issue 300.

This is a more convenient/standard way to get actual values that
should be procs. You can still manually pass a proc in as a
proc actual value OR you can use the block slot to pass in a proc
actual value implicitly.

This also added `Assert::ActualValue.not_given` to use to know
whether a non-block-slot actual value was given or not. I can't do
`nil` or "truthy" checks as these actual values could be anything.
Using the "not given" pattern ensures this will work with any
value.